==> analysis/01-cpue-sdmTMB.Rmd <==
---
title: "Bottom drawl dogfish commerical CPUE"
author: "`r params$author`"
date: "`r Sys.Date()`"
output: html_document
params:
   author: "Sean Anderson"
   species_proper: "SPINY DOGFISH"
   area: !r c("^5A|^5B|^5C|^5D|^5E|^3C|^3D")
   area_name: !r c("Coastwide")
   skip_single_variable_models: FALSE
   era: "modern"
   use_alt_year: FALSE
   alt_year_start_date: "02-21"
   min_year_historic: 1956
   discard_only: FALSE
   parallel: FALSE
   min_positive_tows: 100
   min_positive_trips: 5
   min_yrs_with_trips: 5
   final_year: 2023
   final_date: "2023-12-31"
   depth_bin_quantiles: !r c(0.001, 0.999)
   lat_range: !r c(48, Inf)
   depth_range: !r c(-Inf, Inf)
---

```{r setup, include=FALSE}
spp <- gsub(" ", "-", gsub("\\/", "-", tolower(params$species_proper)))
knitr::opts_chunk$set(
  collapse = TRUE,
  warning = FALSE,
  message = FALSE,
  comment = "#>",
  fig.asp = 0.618,
  cache.path = paste0(spp, "-", params$era, "-", params$area_name, "-cache/"),
  fig.path = paste0(spp, "-", params$era,  "-", params$area_name, "-fig/"),
  echo = TRUE,
  autodep = TRUE,
  cache = TRUE,
  cache.comments = FALSE
)
```

```{r pkgs, cache=FALSE, warning=FALSE, message=FALSE}
library("dplyr")
library("ggplot2")
library("sdmTMB")
library(here)
ggplot2::theme_set(gfplot::theme_pbs())
dir.create(here("data"), showWarnings = FALSE)
dir.create(here("data", "generated"), showWarnings = FALSE)

# Set French language option
FRENCH <- TRUE

# Create appropriate figure directories
if (FRENCH) {
  dir.create("figs-french", showWarnings = FALSE)
  dir.create("figs-french/cpue", showWarnings = FALSE)
  fig_dir <- "figs-french"
} else {
  dir.create("figs/cpue", showWarnings = FALSE)
  fig_dir <- "figs"
}

# Translation helper function
tr <- function(english, french) {
  if (FRENCH) french else english
}

# Helper function for figure paths
fig_path <- function(filename) {
  file.path(fig_dir, filename)
}
```

```{r cpue-params-print}
print(params)
```

```{r cpue}
if (params$era == "modern") {
  fi <- here::here("../cpue-reports/data/cpue-modern.rds")
  if (!file.exists(fi)) {
    d1996 <- gfdata::get_cpue_index(gear = "bottom trawl", min_cpue_year = 1996)
    write_rds(d1996, fi)
  } else {
    d1996 <- readr::read_rds(fi)
  }
  d1996$fishing_event_id_unique <- paste0(d1996$vessel_registration_number, "-", d1996$trip_id, "-", d1996$fishing_event_id)
} else {
  fi <- here::here("../cpue-reports/data/cpue-historical.rds")
  if (!file.exists(fi)) {
    d <- gfdata::get_cpue_historical(species = NULL, end_year = 1995, 
      alt_year_start_date = params$alt_year_start_date)
    write_rds(d, fi)
  } else {
    d <- readr::read_rds(fi)
  }
}
```

Define our fleet. See the help for `?gfplot::tidy_cpue_index` and `?gfplot::tidy_cpue_historical` to see a definition of these arguments. <https://github.com/pbs-assess/gfplot>

```{r define-fleet}
devtools::load_all(here::here("../gfplot"))
dat <- gfplot::tidy_cpue_index(
  d1996,
  species_common = tolower(params$species_proper),
  gear = "bottom trawl",
  alt_year_start_date = params$alt_year_start_date,
  use_alt_year = params$use_alt_year,
  year_range = c(1996, params$final_year),
  lat_range = params$lat_range,
  min_positive_tows = params$min_positive_tows,
  min_positive_trips = params$min_positive_trips,
  min_yrs_with_trips = params$min_yrs_with_trips,
  depth_band_width = 1,
  area_grep_pattern = params$area,
  depth_bin_quantiles = params$depth_bin_quantiles,
  min_bin_prop = 0.001,
  lat_band_width = 0.02,
  return_raw_data = TRUE #<
)
dat$area <- params$area_name
glimpse(dat)
```


```{r}
library(sdmTMB)
dat <- add_utm_columns(dat, c("longitude", "latitude"))
dat$log_depth <- log(dat$best_depth)
dat$vessel <- as.factor(dat$vessel_registration_number)
dat$month <- factor(dat$month)
dat <- filter(dat, year <= 2023)
filter(dat, hours_fished > 2000) # 8762.05!
dat <- filter(dat, hours_fished < 2000)
dat$depth_scaled <- (dat$log_depth - mean(dat$log_depth)) / sd(dat$log_depth)
```

```{r}
ggplot(dat, aes(as.factor(year), (spp_catch + 1)/ hours_fished)) + 
  geom_boxplot() +
  scale_y_log10()
```

```{r}
ggplot(dat, aes(as.factor(year), vessel, size = log(spp_catch / hours_fished))) + 
  geom_point()
```

```{r}
mesh <- make_mesh(dat, c("X", "Y"), cutoff = 30)
plot(mesh$mesh)
```

```{r}
grid <- gfplot::synoptic_grid
grid$log_depth <- log(grid$depth)
grid$survey_domain_year <- NULL
gg <- replicate_df(grid, "year", time_values = sort(unique(dat$year)))
gg$utm_zone <- NULL
gg$cell_area <- NULL
gg$survey_series_name <- NULL
gg$month <- factor(6, levels = levels(dat$month))
gg$vessel <- NA
gg$depth_scaled <- (gg$log_depth - mean(dat$log_depth)) / sd(dat$log_depth)
```

# delta lognormal fit

```{r}
hist(dat$spp_catch)
hist(log(dat$spp_catch + 1))
hist(filter(dat, spp_catch > 0) |> pull(spp_catch))
hist(filter(dat, spp_catch > 0) |> pull(spp_catch) |> log())
```

```{r}
fit_sp <- sdmTMB(
  spp_catch ~ 0 + as.factor(year) + 
    depth_scaled + I(depth_scaled^2) + 
    (1 | vessel) + as.factor(month),
  family = delta_lognormal(),
  mesh = mesh,
  offset = log(dat$hours_fished),
  spatial = "on",
  spatiotemporal = "off",
  data = dat,
  time = "year",
  anisotropy = TRUE,
  # priors = sdmTMBpriors(
  #   matern_s = pc_matern(range_gt = 40, sigma_lt = 3)
  # ),
  predict_args = list(newdata = gg, re_form_iid = NA),
  index_args = list(area = rep(4, nrow(gg))),
  do_index = TRUE,
  silent = FALSE
)

# # heavier than lognormal!? see histogram above and QQ plots below
# fit_dgm <- update(
#   fit_sp, 
#   family = delta_lognormal_mix(), 
#   control = sdmTMBcontrol(
#     start = list(logit_p_mix = qlogis(0.01)),
#     map = list(logit_p_mix = factor(NA))
#   )
# )
# ind_dgm <- get_index(fit_dgm, bias_correct = F)
# ggplot(ind_sp, aes(year, ymin = lwr, ymax = upr, y = est)) + 
#   geom_pointrange()

# AIC(fit_dgm, fit_sp)
# ss <- simulate(fit_dgm, 50)
# sdmTMBextra::dharma_residuals(ss, fit_dgm)

# rrr <- residuals(fit_dgm, model = 2)
# .r2 <- rrr[!is.na(rrr)] # zeros!
# .r2 <- .r2[is.finite(.r2)] # zeros!
# qqnorm(as.numeric(.r2));qqline(as.numeric(.r2))
```

```{r}
sanity(fit_sp)
fit_sp
plot_anisotropy(fit_sp)
```

```{r}
ind_sp <- get_index(fit_sp, bias_correct = TRUE)

ggplot(ind_sp, aes(year, ymin = lwr, ymax = upr, y = est)) + 
  geom_pointrange()
```

# lognormal spatiotemporal?

```{r}
tictoc::tic()
fit_spt <- update(
  fit_sp,
  anisotropy = FALSE, # convergence issues if TRUE
  spatiotemporal = "iid"
)
tictoc::toc()
sanity(fit_spt)
fit_spt
fit_spt$sd_report

AIC(fit_sp, fit_spt)

tictoc::tic()
ind_spt <- get_index(fit_spt, bias_correct = TRUE)
tictoc::toc()

ggplot(ind_spt, aes(year, ymin = lwr, ymax = upr, y = est)) + 
  geom_pointrange()

# compare
bind_rows(
  mutate(ind_sp, type = "dl sp"),
  mutate(ind_spt, type = "dl sp st")
) |> 
  group_by(type) |> 
  mutate(
    lwr = lwr / max(est),
    upr = upr / max(est),
    est = est / max(est)
  ) |> 
  ggplot(aes(year, ymin = lwr, ymax = upr, y = est, colour = type)) + 
  geom_pointrange(position = position_dodge(width = 0.5)) +
  coord_cartesian(ylim = c(0, NA), expand = FALSE)

# on its own:

if (FRENCH) options(OutDec = ",")
ind_spt |>
  mutate(
    upr = upr / exp(mean(log(est))),
    lwr = lwr / exp(mean(log(est))),
    est = est / exp(mean(log(est)))
  ) |> 
  ggplot(aes(year, est)) +
  geom_pointrange(aes(ymin = lwr, ymax = upr)) +
  ylab(tr("Standardized commercial trawl CPUE", "CPUE commerciale chalutière standardisée")) + 
  xlab(tr("Year", "Année")) +
  coord_cartesian(
    expand = FALSE, 
    ylim = c(0, 3),
    xlim = c(range(ind_spt$year) + c(-0.5, 0.5))
  ) +
  scale_x_continuous(breaks = seq(1996, 2024, 2))

saveRDS(ind_sp, "data/generated/geostat-sp-ind-cpue.rds")
saveRDS(ind_spt, "data/generated/geostat-spt-ind-cpue.rds")

ggsave(fig_path("cpue/cpue-index-spatiotemporal.png"), width = 6, height = 4)
```

```{r}
p <- predict(fit_spt, newdata = gg, re_form_iid = NA)
source(here::here("analysis/functions/make_sdmTMB_delta_maps.R"))
make_sdmTMB_delta_maps(p, folder = file.path(fig_dir, "cpue"), french = FRENCH)
```

```{r}
make_re_plot <- function(fit, model = 1) {
  tidy(fit, "ran_vals", conf.int = TRUE, model = model) |> 
    mutate(term2 = paste0("vessel ", stringr::str_pad(as.numeric(as.factor(level_ids)), 2, pad = "0"))) |> 
    ggplot(aes(estimate, xmin = conf.low, xmax = conf.high, y = term2)) +
    geom_pointrange() +
    ylab(tr("Vessel", "Navire")) + xlab(tr("Random intercept value", "Valeur d'ordonnée à l'origine aléatoire"))
}
g1 <- make_re_plot(fit_spt, 1) + ggtitle(tr("Encounter probability (logit link)", "Probabilité de rencontre (lien logit)"))
g2 <- make_re_plot(fit_spt, 2) + ggtitle(tr("Positive density (log link)", "Densité positive (lien log)"))
cowplot::plot_grid(g1, g2, ncol = 2)
ggsave(fig_path("cpue/vessel-effects.png"), width = 8, height = 8)

make_month_plot <- function(fit, model = 1) {
  tidy(fit, conf.int = TRUE, model = model) |> 
    filter(grepl("month", term)) |> 
    mutate(month = gsub("as\\.factor\\(month\\)", "", term)) |> 
    mutate(month = stringr::str_pad(month, 2, pad = "0")) |> 
    ggplot(aes(estimate, xmin = conf.low, xmax = conf.high, y = month)) +
    geom_pointrange() +
    ylab(tr("Month effect relative to January", "Effet du mois par rapport à janvier")) + 
    xlab(tr("Coefficient value", "Valeur du coefficient"))
}
g1 <- make_month_plot(fit_spt, 1) + ggtitle(tr("Encounter probability (logit link)", "Probabilité de rencontre (lien logit)"))
g2 <- make_month_plot(fit_spt, 2) + ggtitle(tr("Positive density (log link)", "Densité positive (lien log)"))
cowplot::plot_grid(g1, g2, ncol = 2)
ggsave(fig_path("cpue/month-effects.png"), width = 7, height = 4)
```

```{r}
# s <- simulate(fit_spt, nsim = 100L)
# sdmTMBextra::dharma_residuals(s, fit_spt)
r1 <- residuals(fit_spt, model = 1)
qqnorm(r1)
r2 <- residuals(fit_spt, model = 2)
.r2 <- r2[!is.na(r2)] # zeros!
.r2 <- .r2[is.finite(.r2)] # zeros!
qqnorm(as.numeric(.r2));qqline(as.numeric(.r2))
```

# shallow only?

```{r}
plot(dat$best_depth, dat$spp_catch);abline(v = 100)
median(dat$best_depth)

dat_shallow <- filter(dat, best_depth < 100)
dat_mid <- filter(dat, best_depth >= 100, best_depth < 200)
dat_deep <- filter(dat, best_depth > 200)

grid_shallow <- filter(gg, depth < 100)
grid_mid <- filter(gg, depth >= 100, depth < 200)
grid_deep <- filter(gg, depth > 200)
# plot(grid_deep$X, grid_deep$Y)
```

```{r}
fit_model <- function(dd, this_grid) {
  this_mesh <- make_mesh(dd, c("X", "Y"), cutoff = 25)
  sdmTMB(
    spp_catch ~ 0 + as.factor(year) + 
      depth_scaled + I(depth_scaled^2) + 
      (1 | vessel) + as.factor(month),
    family = delta_lognormal(),
    mesh = this_mesh,
    offset = log(dd$hours_fished),
    spatial = "on",
    anisotropy = TRUE,
    spatiotemporal = "off",
    data = dd,
    time = "year",
    # priors = sdmTMBpriors(
    #   matern_s = pc_matern(range_gt = 40, sigma_lt = 3)
    # ),
    predict_args = list(newdata = this_grid, re_form_iid = NA),
    index_args = list(area = rep(4, nrow(this_grid))),
    do_index = TRUE,
    silent = FALSE
  )
}

ms <- fit_model(dat_shallow, grid_shallow)
sanity(ms)
mm <- fit_model(dat_mid, grid_mid)
md <- fit_model(dat_deep, grid_deep)
sanity(md)
```

```{r}
is <- get_index(ms, bias_correct = TRUE)
im <- get_index(mm, bias_correct = TRUE)
id <- get_index(md, bias_correct = TRUE)

bind_rows(
  mutate(is, type = if(FRENCH) "(a) Peu profond (< 100 m)" else "(a) Shallow (< 100 m)"),
  mutate(im, type = if(FRENCH) "(b) Moyen (>= 100 m, < 200 m)" else "(b) Mid (>= 100 m, < 200 m)"),
  mutate(id, type = if(FRENCH) "(c) Profond (> 200 m)" else "(c) Deep (> 200 m)")
) |> 
  group_by(type) |> 
  mutate(geo_mean = exp(mean(log(est)))) |> 
  mutate(
    lwr = lwr / geo_mean,
    upr = upr / geo_mean,
    est = est / geo_mean
  ) |> 
  ggplot(aes(year, ymin = lwr, ymax = upr, y = est, group = type)) + 
  geom_pointrange(position = position_dodge(width = 0.5)) +
  coord_cartesian(ylim = c(0, NA), expand = FALSE, xlim = range(is$year) + c(-0.5, 0.5)) +
  facet_wrap(~type) + ylab(tr("Standardized CPUE", "CPUE standardisée")) + xlab(tr("Year", "Année")) +
  geom_smooth(se = FALSE, colour = "red", method = "glm", 
    method.args = list(family = Gamma(link = "log")))
ggsave(fig_path("cpue/cpue-by-depth.png"), width = 9.5, height = 3.5)
```

# what about month of year

```{r}
fit_model_season <- function(months) {
  
  dd <- filter(dat, month %in% months)
  this_grid <- filter(gg, year %in% unique(dd$year))
  this_grid$month <- min(months)
  this_mesh <- make_mesh(dd, c("X", "Y"), cutoff = 25)
  
  sdmTMB(
    spp_catch ~ 0 + as.factor(year) + 
      depth_scaled + I(depth_scaled^2) + 
      (1 | vessel),
    family = delta_lognormal(),
    mesh = this_mesh,
    offset = log(dd$hours_fished),
    spatial = "on",
    spatiotemporal = "off",
    data = dd,
    time = "year",
    anisotropy = TRUE,
    # priors = sdmTMBpriors(
    #   matern_s = pc_matern(range_gt = 40, sigma_lt = 3)
    # ),
    predict_args = list(newdata = this_grid, re_form_iid = NA),
    index_args = list(area = rep(4, nrow(this_grid))),
    do_index = TRUE,
    silent = FALSE
  )
}

library(future)
plan(multisession, workers = 6)
fits <- furrr::future_map(1:12, fit_model_season)
plan(sequential)

fits2 <- fits
names(fits2) <- stringr::str_pad(1:12, 2, pad = "0")

plan(multisession, workers = 4)
i2 <- furrr::future_map_dfr(fits2, get_index, bias_correct = TRUE, .id = "month")
plan(sequential)

i2 |> ggplot(aes(year, ymin = lwr, ymax = upr, y = est)) + 
  geom_pointrange() +
  coord_cartesian(ylim = c(0, NA), expand = FALSE) +
  facet_wrap(~month)

i2 |> ggplot(aes(year, ymin = lwr/100000, ymax = upr/100000, y = est/100000, colour = month, fill = month)) + 
  geom_line() +
  geom_ribbon(alpha = 0.2, colour = NA) +
  coord_cartesian(ylim = c(0, NA), expand = FALSE) +
  scale_fill_viridis_d() +
  scale_colour_viridis_d() + ylab(tr("Standardized CPUE", "CPUE standardisée")) + xlab(tr("Year", "Année")) + 
  labs(colour = tr("Month", "Mois"), fill = tr("Month", "Mois")) +
  theme(axis.text.y.left = element_blank(), axis.ticks.y.left = element_blank())
ggsave(fig_path("cpue/cpue-lines-by-month-abs.png"), width = 5, height = 3.8)

i2 |>
  ggplot(aes(year, ymin = lwr/100000, ymax = upr/100000, y = est/100000)) + 
  geom_line() +
  geom_ribbon(alpha = 0.2, colour = NA) +
  coord_cartesian(ylim = c(0, NA), expand = FALSE) +
  scale_fill_viridis_d() +
  scale_colour_viridis_d() +
  facet_wrap(~month) +
  geom_smooth(se = F, colour = "red", method = "glm", 
    method.args = list(family = Gamma(link = "log"))) +
  ylab(tr("Standardized CPUE", "CPUE standardisée")) + xlab(tr("Year", "Année")) +
  theme(axis.text.y.left = element_blank(), axis.ticks.y.left = element_blank())
ggsave(fig_path("cpue/cpue-lines-by-month-facets-abs.png"), width = 8, height = 6)
  
i2 |> group_by(month) |> 
  mutate(geo_mean = exp(mean(log(est)))) |> 
  mutate(
    lwr = lwr / geo_mean,
    upr = upr / geo_mean,
    est = est / geo_mean
  ) |> 
  ggplot(aes(year, ymin = lwr, ymax = upr, y = est)) + 
  geom_line() +
  geom_ribbon(alpha = 0.2, colour = NA) +
  coord_cartesian(ylim = c(0, 4), expand = FALSE) +
  scale_fill_viridis_d() +
  scale_colour_viridis_d() +
  facet_wrap(~month) +
  geom_smooth(se = FALSE, colour = "red", method = "glm", 
    method.args = list(family = Gamma(link = "log"))) +
  ylab(tr("Standardized CPUE", "CPUE standardisée")) + xlab(tr("Year", "Année"))
ggsave(fig_path("cpue/cpue-lines-by-month-facets.png"), width = 8, height = 6)

i2 |> group_by(month) |> 
  mutate(geo_mean = exp(mean(log(est)))) |> 
  mutate(
    lwr = lwr / geo_mean,
    upr = upr / geo_mean,
    est = est / geo_mean
  ) |> 
  ggplot(aes(year, ymin = lwr, ymax = upr, y = est, 
    colour = month, fill = month, group = month)) + 
  geom_line(alpha = 0.7) +
  geom_ribbon(alpha = 0.1, colour = NA) +
  coord_cartesian(ylim = c(0, 4), expand = FALSE) +
  scale_fill_viridis_d() +
  scale_colour_viridis_d() +
  geom_smooth(se = F, method = "glm", method.args = list(family = Gamma(link = "log"))) +
  ylab(tr("Standardized CPUE", "CPUE standardisée")) + xlab(tr("Year", "Année")) + 
  labs(colour = tr("Month", "Mois"), fill = tr("Month", "Mois"))
ggsave(fig_path("cpue/cpue-lines-by-month.png"), width = 5, height = 3.8)

i2 |> group_by(month) |> 
  group_split() |> 
  purrr::map_dfr(function(.x) {
    .x$year <- .x$year / 10
    mm <- glm(est ~ year, family = Gamma(link = "log"), data = .x)
    data.frame(month = unique(.x$month), slope = coef(mm)[[2]], 
      lwr = confint(mm)[2,1], upr = confint(mm)[2,2])
  }) |> 
  ggplot(aes(month, 100*(1 - exp(slope)), ymin = 100*(1 - exp(lwr)), ymax = 100*(1 - exp(upr)))) + 
  geom_pointrange() + 
  ylab(tr("% decline per decade", "% de déclin par décennie")) + xlab(tr("Month", "Mois"))
ggsave(fig_path("cpue/cpue-slopes-by-month.png"), width = 5, height = 3.8)
if (FRENCH) options(OutDec = ".")
```
